+++ src/ui/components/Filters/filter.module.scss
@@ -1,0 +1,18 @@
+@import '../../globalStyles.scss';
+
+.datePickerField {
+  height: 4rem;
+  background-color: $whiteColor;
+  border: 1px solid var(--lightGrey);
+  border-radius: 4px;
+  box-sizing: border-box;
+  display: flex;
+  // width: 93%;
+  font-size: 1.6rem;
+  color: $blackColor;
+  padding-left: 1.6rem;
+  padding-right: 1.6rem;
+  outline: 0;
+  margin-top: 26px;
+  line-height: 37px;
+}
