+++ src/ui/layouts/stacks/Stacks/WorkspaceDropdown.module.scss
@@ -4,23 +4,6 @@ $activeBgColor: #fbfbfb;
 $boxShadow: 0px 2px 4px rgba(0, 0, 0, 0.05);
 $width: 18rem;
 
-.datePickerField {
-  height: 4rem;
-  background-color: $whiteColor;
-  border: 1px solid var(--lightGrey);
-  border-radius: 4px;
-  box-sizing: border-box;
-  display: flex;
-  // width: 93%;
-  font-size: 1.6rem;
-  color: $blackColor;
-  padding-left: 1.6rem;
-  padding-right: 1.6rem;
-  outline: 0;
-  margin-top: 26px;
-  line-height: 37px;
-}
-
 .dropdown {
   background-color: $whiteColor;
   position: absolute;
